+++ eventmesh-runtime/src/test/java/org/apache/eventmesh/runtime/client/api/PubClient.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.eventmesh.runtime.client;
+package org.apache.eventmesh.runtime.client.api;
 
 import org.apache.eventmesh.common.protocol.tcp.Package;
 import org.apache.eventmesh.common.protocol.tcp.UserAgent;
