+++ eventmesh-runtime/src/test/java/org/apache/eventmesh/runtime/client/impl/EventMeshClientImpl.java
@@ -22,9 +22,9 @@ import org.apache.eventmesh.common.protocol.tcp.Package;
 import org.apache.eventmesh.common.protocol.SubscriptionMode;
 import org.apache.eventmesh.common.protocol.tcp.UserAgent;
 
-import org.apache.eventmesh.runtime.client.EventMeshClient;
-import org.apache.eventmesh.runtime.client.PubClient;
-import org.apache.eventmesh.runtime.client.SubClient;
+import org.apache.eventmesh.runtime.client.api.EventMeshClient;
+import org.apache.eventmesh.runtime.client.api.PubClient;
+import org.apache.eventmesh.runtime.client.api.SubClient;
 import org.apache.eventmesh.runtime.client.common.UserAgentUtils;
 import org.apache.eventmesh.runtime.client.hook.ReceiveMsgHook;
 
@@ -78,21 +78,21 @@ public class EventMeshClientImpl implements EventMeshClient {
         this.subClient.heartbeat();
     }
 
-
     public Package listen() throws Exception {
         return this.subClient.listen();
     }
 
     @Override
-    public Package justSubscribe(String topic, SubscriptionMode subscriptionMode, SubcriptionType subcriptionType) throws Exception {
+    public Package justSubscribe(String topic, SubscriptionMode subscriptionMode,
+        SubcriptionType subcriptionType) throws Exception {
         return this.subClient.justSubscribe(topic, subscriptionMode, subcriptionType);
     }
 
     @Override
-    public Package justUnsubscribe(String topic, SubscriptionMode subscriptionMode, SubcriptionType subcriptionType) throws Exception {
+    public Package justUnsubscribe(String topic, SubscriptionMode subscriptionMode,
+        SubcriptionType subcriptionType) throws Exception {
         return this.subClient.justUnsubscribe(topic, subscriptionMode, subcriptionType);
     }
-
 
     public void registerSubBusiHandler(ReceiveMsgHook handler) throws Exception {
         this.subClient.registerBusiHandler(handler);
@@ -105,10 +105,10 @@ public class EventMeshClientImpl implements EventMeshClient {
     @Override
     public String toString() {
         return "AccessClientImpl{" +
-                "accessHost='" + accessHost + '\'' +
-                ", accessPort=" + accessPort +
-                ", agent=" + agent +
-                '}';
+            "accessHost='" + accessHost + '\'' +
+            ", accessPort=" + accessPort +
+            ", agent=" + agent +
+            '}';
     }
 
     @Deprecated
